sentinel/requirements.txt
kubernetes>=28.0.0
google-generativeai>=0.8.0
requests>=2.31.0
